Rename record serializers and deserializers
